[Silver II] Title: 관리 난항, Time: 432 ms, Memory: 42856 KB -BaekjoonHub

+++ 백준/Silver/9367. 관리 난항/관리 난항.java
@@ -35,7 +35,7 @@ public class Main {
 
         void processAccident(double damageCost) {
             // 수리비는 올림 처리
-            this.totalCost += Math.ceil(damageCost);
+            this.totalCost += (int) Math.ceil(damageCost);
         }
     }
